Add `(entries: true)` argument support
This allows to iterate over keys of a javascript object treating it like
a dict.

I.e. filter like ``{ field(entries: true) }`` is treated like the data
in the field not an object of ``{"key1": "value1"}`` but an array of
``[{"key": "key1", "value": "value1"}]``.

In other words, this allows working with javascript objects with keys
unknown in advance.